Added IDRule to the GUI
refs #5073

Added views, forms and template for displaying and adding IDRules.  When
saved, IDRule takes an optional parameter of what object it is replacing.  ID
Rule added to the GUI, both top bar and sidebar.  django-annoying added as a
dependency.

IDCommand and IDTool are now connected by a many to many relationship through
IDToolConfig.  This was semantically true before, and is not reflected in the
code.