IDCommand stores entire script
refs #5440

IDCommand stores the entire command to be run and what type of command
it is (bash, python, command line, stores own shebang).  In
Archivematica, scripts with a type of bash or python will have the
appropriate shebang prepended, while as-is will not.  A description
field allows human-friendly display of the script.