Normalized names, added revision page.
refs #5531

Added the revisions page in rough.

Normalized names of views and URLs so they correspond to the MySQL table
names. This way the model can be used to work out the name of views.
This is needed as the revisions page is given a name corresponding to a
model and ideally can work, generically, from that.