Tweak to support revision reversion.
refs #5531

Added a minor tweak to streamline using old versions to create new versions.
If an old version (with the "enabled" field set to "false") is edited, view logic can
pull the most recent ancestor and assign that to the "replacing" kwarg during
a save of the old version.

The old version will then be made into a new version by setting its pk to "None"
(as we normally do) and becase the version was made from an old version we need to
also set "enabled" to "True".