FPRule.output_format is foreign key, IDToolConfig updates MCP properly
refs #5627, #5653

FPRule.output_file_format now FPRule.output_format, and is a foreign key to
FormatVersion, which should be the format that the command produces. #5627

Updated IDToolConfig.save so that when it saves, it inserts its own UUID
and description into MicroServiceChoiceReplacementDic.  This fixes #5653
and avoids problems where the IDCommand is updated but the
MicroServiceChoiceReplacementDic is pointing at the wrong row.  Requires
the lookup in the MCP.

Also updated VersionedModel.replaces to point at the uuid field.