Removed many-to-many relationships between tools and commands
refs #5272

Removed IDToolConfig and FPCommandTool, since they were causing a variety
of problems, and a many-to-one relationship was sufficient.

IDToolConfig replaced with a foreign key from IDCommand to IDTool, and
config was moved to IDCommand.  FPCommandTool replaced with a foreign key
from FPCommand to FPTool.  Models, forms and views updated for the change.