Pass valid message arg to ValidationError
ValidationError expects a dictionary that maps field names to lists of errors,
before this commit we were passing tuples instead of lists.

From @jrwdhunham via #62.

For more clarification on the rationale behind this, observe that passing tuples
instead of lists results in Django turning the tuple into a string and putting
it in a list:

```
try:
    raise ValidationError({'__all__': ('abc', 'def')})
except ValidationError as error:
    print(error)
{'__all__': ["('abc', 'def')"]}

try:
    raise ValidationError({'__all__': ['abc', 'def']})
except ValidationError as error:
    print(error)
{'__all__': ['abc', 'def']}
```