Travis: run tests against MySQL and SQLite
Our initial migration test only covered SQLite. We found later that AM17
deployments making use of MySQL in strict mode (the default in recent releases)
failed:

    django.db.utils.DataError: Problem installing fixture '/home/travis/build/artefactual/archivematica-fpr-admin/fpr/migrations/pronom_92.json': Could not load fpr.Format(pk=None): (1406, "Data too long for column 'slug' at row 1")

Django's CharField validator doesn't seem to kick in during the migration - I
suspect this is not a bug in Django but it django-autoslug's AutoSlugField
which subclasses the former.